feat(transaction): add domain transaction with mapper

+++ bankAccount-domain/src/main/java/fr/kata/bankAccount/domain/BankAccountService.java
@@ -3,13 +3,15 @@ package fr.kata.bankAccount.domain;
 import fr.kata.bankAccount.domain.data.Account;
 import fr.kata.bankAccount.domain.data.Bank;
 
+import java.math.BigDecimal;
+
 public interface BankAccountService {
 
     Account getAccount(Long ownerId, Long accountId);
 
-    Account deposit(Long accountId, Long ownerId, Float amount);
+    Account deposit(Long accountId, Long ownerId, BigDecimal amount);
 
-    Account withdraw(Long accountId, Long ownerId, Float amount);
+    Account withdraw(Long accountId, Long ownerId, BigDecimal amount);
 
     boolean isOwnerBank(Bank ownerBank);
 }
